Splitting multiple classes/interfaces into separate files.
